+++ godot-core/src/tools/gfile.rs
@@ -8,9 +8,9 @@
 use crate::builtin::{real, GString, PackedByteArray, PackedStringArray, Variant};
 use crate::classes::file_access::{CompressionMode, ModeFlags};
 use crate::classes::FileAccess;
-use crate::extras::IoError;
 use crate::global::Error;
 use crate::obj::Gd;
+use crate::tools::IoError;
 
 use std::cmp;
 use std::io::{BufRead, ErrorKind, Read, Seek, SeekFrom, Write};
@@ -50,7 +50,7 @@ use std::io::{BufRead, ErrorKind, Read, Seek, SeekFrom, Write};
 /// ```no_run
 /// use godot::builtin::GString;
 /// use godot::classes::file_access::ModeFlags;
-/// use godot::extras::GFile;
+/// use godot::tools::GFile;
 ///
 /// fn save_game() -> Result<(), std::io::Error> {
 ///
