+++ godot-core/src/tools/save_load.rs
@@ -7,9 +7,9 @@
 
 use crate::builtin::GString;
 use crate::classes::{Resource, ResourceLoader, ResourceSaver};
-use crate::extras::IoError;
 use crate::global::Error as GodotError;
 use crate::obj::{Gd, Inherits};
+use crate::tools::IoError;
 
 /// ⚠️ Loads a resource from the filesystem located at `path`, panicking on error.
 ///
