+++ itest/rust/src/builtin_tests/script/script_instance_tests.rs
@@ -10,11 +10,11 @@ use std::ffi::c_void;
 use godot::builtin::meta::{ClassName, FromGodot, MethodInfo, PropertyInfo, ToGodot};
 use godot::builtin::{GString, StringName, Variant, VariantType};
 use godot::classes::{IScriptExtension, Object, Script, ScriptExtension, ScriptLanguage};
-use godot::extras::{create_script_instance, ScriptInstance, SiMut};
 use godot::global::{MethodFlags, PropertyHint, PropertyUsageFlags};
 use godot::obj::{Base, Gd, WithBaseField};
 use godot::register::{godot_api, GodotClass};
 use godot::sys;
+use godot::tools::{create_script_instance, ScriptInstance, SiMut};
 
 #[derive(GodotClass)]
 #[class(base = ScriptExtension, init)]
