+++ itest/rust/src/engine_tests/translate_test.rs
@@ -7,7 +7,7 @@
 
 use crate::framework::itest;
 use godot::builtin::Vector2;
-use godot::extras::{tr, tr_n};
+use godot::tools::{tr, tr_n};
 
 #[itest]
 fn tr_macro_format() {
